% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ex_biopax}
\alias{ex_biopax}
\title{Biopax example data}
\description{
A dataset containing Porphyrin metabolism pathway in Biopax Level 3 and parsed with
\code{\link[rBiopaxParser]{readBiopax}}.
}
\examples{
data(ex_biopax)
ex_biopax
}

