% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPM-package.R
\docType{data}
\name{ex_kgml_sig}
\alias{ex_kgml_sig}
\title{Singaling network from KGML example}
\description{
An example igraph object representing Ras and chemokine signaling pathways in human 
extracted from KGML files.
}
\examples{
data(ex_kgml_sig)
plotNetwork(ex_kgml_sig, vertex.color="pathway")

}

