% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPM-package.R
\docType{data}
\name{ex_microarray}
\alias{ex_microarray}
\title{An microarray data example.}
\description{
An microarray data example. This is part of the ALL dataset, for demonstration purposes.
}
\examples{
data(ex_microarray)

}

