% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPM-package.R
\name{registerMemoryErr}
\alias{registerMemoryErr}
\title{Internal method to register memery errors.}
\usage{
registerMemoryErr(method)
}
\arguments{
\item{method}{The mathod which generated the error.}
}
\description{
Internal method to register memery errors, caused by compiled code. This method
is used only by the package, and should not be invoked by users.
}
\author{
Ahmed Mohamed
}
