% Generated by roxygen2 (4.0.1): do not edit by hand
\name{samplePaths}
\alias{samplePaths}
\title{Creates a set of sample path p-values for each length given a weighted network}
\usage{
samplePaths(graph, max.path.length, num.samples = 1000, num.warmup = 10,
  verbose = TRUE)
}
\arguments{
\item{graph}{A weighted igraph object. Weights must be in \code{edge.weights} or \code{weight}
edge attributes.}

\item{max.path.length}{The maxmimum path length.}

\item{num.samples}{The numner of paths to sample}

\item{num.warmup}{The number of warm up paths to sample.}

\item{verbose}{Whether to display the progress of the function.}
}
\value{
A matrix where each row is a path length and each column is the number of paths sampled.
}
\description{
Randomly traverses paths of increasing lengths within a set network to create an
empirical pathway distribution for more accurate determination of path significance.
}
\details{
Can take a bit of time.
}
\examples{
## Prepare a weighted reaction network.
	## Conver a metabolic network to a reaction network.
 data(ex_sbml) # bipartite metabolic network of Carbohydrate metabolism.
 rgraph <- makeReactionNetwork(ex_sbml, simplify=TRUE)

	## Assign edge weights based on Affymetrix attributes and microarray dataset.
 # Calculate Pearson's correlation.
	data(ex_microarray)	# Part of ALL dataset.
	rgraph <- assignEdgeWeights(microarray = ex_microarray, graph = rgraph,
		weight.method = "cor", use.attr="miriam.uniprot",
		y=factor(colnames(ex_microarray)), bootstrap = FALSE)

	## Get significantly correlated paths using "p-valvue" method.
	##   First, establish path score distribution by calling "samplePaths"
 pathsample <- samplePaths(rgraph, max.path.length=10,
                        num.samples=100, num.warmup=10)

	##   Get all significant paths with p<0.1
	significant.p <- pathRanker(rgraph, method = "pvalue",
                sampledpaths = pathsample ,alpha=0.1)
}
\author{
Timothy Hancock

Ahmed Mohamed
}
\seealso{
Other Path ranking methods: \code{\link{extractPathNetwork}};
  \code{\link{getPathsAsEIDs}}; \code{\link{pathRanker}}
}

