% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPM-package.R
\docType{data}
\name{ex_biopax}
\alias{ex_biopax}
\title{Biopax example data}
\description{
A dataset containing Porphyrin metabolism pathway in Biopax Level 3 and parsed with
\code{\link[rBiopaxParser]{readBiopax}}.
}
\examples{
data(ex_biopax)
ex_biopax

}
