% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ex_kgml_sig}
\alias{ex_kgml_sig}
\title{Singaling network from KGML example}
\description{
An example igraph object representing Ras and chemokine signaling pathways in human
extracted from KGML files.
}
\examples{
data(ex_kgml_sig)
plotNetwork(ex_kgml_sig, vertex.color="pathway")
}

